+++ app/src/main/kotlin/vision/genesis/android/mvp/views/EnterAmountView.kt
@@ -6,5 +6,5 @@ import vision.genesis.android.mvp.models.data.TraderInfo
 
 interface EnterAmountView : MvpView {
     fun showTraderInfo(traderInfo: TraderInfo)
-    fun showUserInfo(availableTokens: Int, bidForToken: Float)
+    fun showUserInfo(availableTokens: Float, bidForToken: Float)
 }
